refactor landing.js page/ install react-router-dom
